move stuff about.

+++ bgg-api/Src/Entities/Dtos/PlayerCountScoreDto.cs
@@ -3,5 +3,5 @@ namespace BGGAPI.Src.Entities.Dtos;
 public record PlayerCountScoreDto
 (
     int PlayerCount,
-    double Score
+    float Score
 );
