+++ bgg-api/Src/Entities/Dtos/RecommendationDto.cs
@@ -2,6 +2,6 @@ namespace BGGAPI.Src.Entities.Dtos;
 
 public record RecommendationDto
 (
-    double Score,
+    float Score,
     BoardGameCardDto BoardGame
 );
